Refactor api gateway into util.
